/ndcs/:code/full => /api/v1/ndcs/:code/full for consistency and to avoid potential conflicts with frontend routes
