+++ config/routes.rb
@@ -2,12 +2,10 @@ Rails.application.routes.draw do
   namespace :api do
     namespace :v1 do
       resources :countries, only: [:index], controller: :countries
+      resources :ndcs, param: :code, only: [:index, :show] do
+        get :full, on: :member, controller: :ndc_full_texts, action: :show
+      end
     end
-  end
-
-  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
-  resources :ndcs, param: :code, only: [:index, :show] do
-    get :full, on: :member, controller: :ndc_full_texts, action: :show
   end
 
   root 'application#index'
