+++ spec/controllers/api/v1/ndc_full_texts_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe NdcFullTextsController, type: :controller do
+RSpec.describe Api::V1::NdcFullTextsController, type: :controller do
   describe 'GET show' do
     let(:poland) { FactoryGirl.create(:location, iso_code3: 'POL') }
     let(:poland_img_src) { 'POL-1.PNG' }
